feat: Add video stories page and update navigation structure
