+++ package.json
@@ -22,6 +22,7 @@
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-icons": "^5.3.0",
+    "react-pdf": "^9.1.1",
     "react-player": "^2.16.0",
     "swiper": "^11.1.14"
   },
