+++ src/pages/topluluk/etkinliklerimiz.module.css
@@ -83,7 +83,7 @@
     overflow: hidden;
     text-overflow: ellipsis;
     white-space: nowrap;
-    max-width: 300px; 
+    max-width: 250px; 
 }
 
 @media (max-width: 1500px) {
@@ -128,6 +128,34 @@
     }
     
     .videoInfo h2 {
-        max-width: 100px; 
+        max-width: 150px; 
     }
+}
+
+.pagination {
+    display: flex;
+    justify-content: center;
+    gap: 0.5rem;
+}
+
+.pageItem {
+    padding: 10px 20px;
+    cursor: pointer;
+    border-radius: 5px;
+    font-weight: 500;
+    transition: background-color 0.3s, box-shadow 0.3s;
+    border: none; 
+    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1); 
+}
+
+.pageItem:hover {
+    color: #e4b902; 
+    background-color: rgb(53, 53, 53);
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2); 
+}
+
+.pageItem.active {
+    background-color: #e4b902;
+    color: #fff; 
+    font-weight: bold; 
 }
